% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-doc.R
\docType{package}
\name{DGQGSI-package}
\alias{DGQGSI}
\alias{DGQGSI-package}
\title{DGQGSI: Desired-Gain Selection Index and Desired-Gain QGSI Tools}
\description{
Package providing tools for:
\itemize{
\item desired-gain selection index optimization,
\item desired-gain adapted QGSI computation from trait GEBVs,
\item automatic construction of quadratic desired-gain matrices,
\item comparison of DG and QGSI outputs,
\item a high-level wrapper for full pipeline execution.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FAkohoue/DGQGSI}
  \item \url{https://fakohoue.github.io/DGQGSI/}
  \item Report bugs at \url{https://github.com/FAkohoue/DGQGSI/issues}
}

}
\author{
\strong{Maintainer}: Félicien Akohoue \email{akohoue.f@gmail.com}

}
\keyword{internal}
